Fix (or more likely hack around) issue with mounting the config file as a directory.

+++ src/Api/Program.cs
@@ -15,7 +15,7 @@ namespace Sharpsonic.Api {
             .ConfigureAppConfiguration((buildercontext, config) => {
                 // TODO: Hardcoded paths - can we pick this up from the environment?
                 // E.g. what happens on a Windows container?
-                if (!File.Exists("/config/appsettings.json")) {
+                if (!File.Exists("/config/appsettings.json") && Directory.Exists("/config")) {
                     File.Copy("appsettings.json", "/config/appsettings.json");
                 }
